Improve Windows Compatibility (for csrc/scripts) (#2941)

+++ torch/csrc/generic/Tensor.cpp
@@ -218,7 +218,7 @@ THTensor* THPTensor_(fromNumpy)(PyObject *numpy_array) {
 
     THLongStoragePtr strides(THLongStorage_newWithSize(ndim));
     int64_t *strides_data = strides->data;
-    long elsize = PyArray_ITEMSIZE(array);
+    int64_t elsize = PyArray_ITEMSIZE(array);
     for (int i = 0; i < ndim; ++i) {
       // numpy uses bytes, torch uses elements
       // we have to cast sizeof to long, because otherwise stride gets
@@ -382,7 +382,7 @@ static PyObject * THPTensor_(pynew)(PyTypeObject *type, PyObject *args, PyObject
           "sequences and there's no way to infer how many dimension should "
           "the tensor have");
       THPUtils_assert(length > 0, "given sequence has an invalid size of "
-          "dimension %ld: %ld", (long)sizes.size(), (long)length);
+          "dimension %" PRId64 ": %" PRId64, (int64_t)sizes.size(), (int64_t)length);
       item = PySequence_GetItem(item, 0);
       if (!item)
         return NULL;
@@ -396,7 +396,7 @@ static PyObject * THPTensor_(pynew)(PyTypeObject *type, PyObject *args, PyObject
       *sizes_data++ = size;
     THTensorPtr tensor(THTensor_(newWithSize)(LIBRARY_STATE sizes_storage, NULL));
 
-    int ndims = sizes.size();
+    int ndims = (int) sizes.size();
     std::vector<size_t> indices(ndims);
     std::vector<THPObjectPtr> sequences(ndims);
     Py_INCREF(first_arg);
@@ -563,7 +563,7 @@ static PyObject * THPTensor_(pynew)(PyTypeObject *type, PyObject *args, PyObject
 #endif
 
 static bool THPTensor_(_indexOnce)(PyObject *index, int &indexed_dim,
-        THTensorPtr &tresult, THStorage* &sresult, long &storage_offset)
+        THTensorPtr &tresult, THStorage* &sresult, int64_t &storage_offset)
 {
 #ifdef WITH_NUMPY
   static PyArray_Descr *NumpyLongArrDescr = PyArray_DescrFromType(NPY_INT64);
@@ -573,7 +573,7 @@ static bool THPTensor_(_indexOnce)(PyObject *index, int &indexed_dim,
   if(IS_SCALAR(index)) {
     int64_t idx;
     UNPACK_SCALAR(index);
-    long dimsize = THTensor_(size)(LIBRARY_STATE tresult.get(), indexed_dim);
+    int64_t dimsize = THTensor_(size)(LIBRARY_STATE tresult.get(), indexed_dim);
 
     // If the user provided negative idx, convert to positive equivalent
     idx = (idx < 0) ? dimsize + idx : idx;
@@ -647,7 +647,7 @@ static bool THPTensor_(_checkSingleSequenceTriggersAdvancedIndexing)(PyObject *a
 }
 
 static bool THPTensor_(_checkBasicIntegerArrayIndexing)(THPTensor *indexed, PyObject *arg) {
-  long ndim = THTensor_(nDimension)(LIBRARY_STATE indexed->cdata);
+  int64_t ndim = THTensor_(nDimension)(LIBRARY_STATE indexed->cdata);
 
   if (PySequence_Check(arg) && PySequence_Size(arg) == ndim) {
     THPObjectPtr fast = THPObjectPtr(PySequence_Fast(arg, NULL));
@@ -687,7 +687,7 @@ static bool THPTensor_(_checkAdvancedIndexing)(THPTensor *indexed, PyObject *arg
   // ':' and/or a single '...', can be less than ndim indexers, all sequencers
   // adjacent
 
-  long ndim = THTensor_(nDimension)(LIBRARY_STATE indexed->cdata);
+  int64_t ndim = THTensor_(nDimension)(LIBRARY_STATE indexed->cdata);
   if (PySequence_Check(arg) && PySequence_Size(arg) <= ndim + 1) {
     THPObjectPtr fast = THPObjectPtr(PySequence_Fast(arg, NULL));
 
@@ -720,7 +720,7 @@ static bool THPTensor_(_checkAdvancedIndexing)(THPTensor *indexed, PyObject *arg
         continue;
       }
       if (PySlice_Check(item)) {
-        long dimSize = THTensor_(size)(LIBRARY_STATE indexed->cdata, correspondingTensorDim);
+        int64_t dimSize = THTensor_(size)(LIBRARY_STATE indexed->cdata, correspondingTensorDim);
         // Basically verify that the Slice is ':' and did not specify
         // a specific start, end or step
         Py_ssize_t start, end, length, step;
@@ -919,7 +919,7 @@ static bool THPTensor_(_convertToTensorIndexers)(
     THLongStorage_set(viewer.get(), 0, nElement);
     for (auto& dimBroadcast : broadcasted) {
       Py_ssize_t dim = dimBroadcast.first;
-      long sizeAtDim = THTensor_(size)(LIBRARY_STATE indexed, dim);
+      int64_t sizeAtDim = THTensor_(size)(LIBRARY_STATE indexed, dim);
 
       // Need to make contiguous to view as 1D :/
       THPPointer<THIndexTensor> contig(THIndexTensor_(newContiguous)(LIBRARY_STATE dimBroadcast.second.get()));
@@ -927,7 +927,7 @@ static bool THPTensor_(_convertToTensorIndexers)(
       // View as 1D + get1D makes me sad :(
       THPPointer<THIndexTensor> flat(THIndexTensor_(newView)(LIBRARY_STATE contig.get(), viewer));
       for (ptrdiff_t i = 0; i < THIndexTensor_(nElement)(LIBRARY_STATE flat.get()); ++i) {
-        long indexAtDim = THTensor_fastGet1d(flat.get(), i);
+        int64_t indexAtDim = THTensor_fastGet1d(flat.get(), i);
         if (indexAtDim >= sizeAtDim) {
           PyErr_Format(PyExc_IndexError, "index %lld from broadcast indexer is out of range "
               "for dimension %lld (of size %lld)",
@@ -968,7 +968,7 @@ static bool THPTensor_(_convertToTensorIndexers)(
   return true;
 }
 
-static inline long THPTensor_(_indexToOffset)(
+static inline int64_t THPTensor_(_indexToOffset)(
     THTensorPtr& indexed,
     std::unordered_map<Py_ssize_t, THPPointer<THIndexTensor>>& broadcasted,
     ptrdiff_t index)
@@ -1030,12 +1030,12 @@ static inline long THPTensor_(_indexToOffset)(
   //
   // Special care needs to be taken to handle advanced indexers at the beginning, end.
 
-  long offset = 0;
-  for (long i = THTensor_(nDimension)(LIBRARY_STATE indexed) - 1; i >= 0; --i) {
+  int64_t offset = 0;
+  for (int64_t i = THTensor_(nDimension)(LIBRARY_STATE indexed) - 1; i >= 0; --i) {
     // Get size at dimension i, its the size of the indexed Tensor at that dimension if its
     // not an advanced indexing dimension, otherwise its the size of the broadcast Tensor
     ptrdiff_t sizeAtDim, indexAtDim, nextIndex;
-    long strideAtDim = THTensor_(stride)(LIBRARY_STATE indexed, i);
+    int64_t strideAtDim = THTensor_(stride)(LIBRARY_STATE indexed, i);
 
     auto broadcast = broadcasted.find(i);
     if (broadcast != broadcasted.end()) {
@@ -1116,7 +1116,7 @@ static THIndexTensor* THPTensor_(_calculateLinearIndices)(
   // Call GPU kernel for index calculation
   THCudaLongTensor *cudaIndices =
     THCudaLongTensor_newWithSize1d(LIBRARY_STATE indexingElements);
-  long baseOffset = THTensor_(storageOffset)(LIBRARY_STATE indexed);
+  int64_t baseOffset = THTensor_(storageOffset)(LIBRARY_STATE indexed);
 
   // Need to pass broadcast Tensors to API, pass NULL ptr for all empty
   // (i.e. not-advanced indexed) dims
@@ -1134,9 +1134,9 @@ static THIndexTensor* THPTensor_(_calculateLinearIndices)(
   return cudaIndices;
 #else
   THIndexTensor *linearIndices = THIndexTensor_(newWithSize1d)(LIBRARY_STATE indexingElements);
-  long baseOffset = THTensor_(storageOffset)(LIBRARY_STATE indexed);
+  int64_t baseOffset = THTensor_(storageOffset)(LIBRARY_STATE indexed);
   for (ptrdiff_t i = 0; i < indexingElements; ++i) {
-    long linearIdx = THPTensor_(_indexToOffset)(
+    int64_t linearIdx = THPTensor_(_indexToOffset)(
         indexed, flattenedBroadcasters, i);
     THTensor_fastSet1d(linearIndices, i, baseOffset + linearIdx);
   }
@@ -1423,7 +1423,7 @@ static PyObject* THPTensor_(advancedIndexSelect)(THPTensor *self, PyObject *args
 
 // Handles indexing into a Tensor given a tuple, ellipses, sequence, etc. index
 static bool THPTensor_(_index)(THPTensor *self, PyObject *index,
-    THTensorPtr &tresult, THStorage * &sresult, long &storage_offset)
+    THTensorPtr &tresult, THStorage * &sresult, int64_t &storage_offset)
 {
   // As a base case, we create a new Tensor that is a copy of the Tensor
   // we are indexing
@@ -1435,10 +1435,10 @@ static bool THPTensor_(_index)(THPTensor *self, PyObject *index,
   if(PyTuple_Check(index)) {
     // num_indexers is the number of indexing objects in the tuple, num_effective_indexers
     // is the number of non-None, non-ellipses indexing objects
-    long num_indexers = (long)PyTuple_Size(index);
-    long num_effective_indexers = num_indexers;
-    long num_tensor_dim = THTensor_(nDimension)(LIBRARY_STATE self->cdata);
-    long ellipsis_pos = -1;
+    int64_t num_indexers = (int64_t)PyTuple_Size(index);
+    int64_t num_effective_indexers = num_indexers;
+    int64_t num_tensor_dim = THTensor_(nDimension)(LIBRARY_STATE self->cdata);
+    int64_t ellipsis_pos = -1;
     for (int i = 0; i < num_indexers; i++) {
       PyObject *indexer = PyTuple_GET_ITEM(index, i);
       if (indexer == Py_Ellipsis) {
@@ -1452,7 +1452,7 @@ static bool THPTensor_(_index)(THPTensor *self, PyObject *index,
     }
     if (num_effective_indexers > num_tensor_dim) {
       PyErr_Format(PyExc_IndexError,
-          "trying to index %ld dimensions of a %ld dimensional tensor",
+          "trying to index %" PRId64 " dimensions of a %" PRId64 " dimensional tensor",
           num_effective_indexers, num_tensor_dim);
       return false;
     }
@@ -1552,7 +1552,7 @@ static PyObject * THPTensor_(getValue)(THPTensor *self, PyObject *index)
 
   THTensorPtr tresult;
   THStorage *sresult;
-  long storage_offset;
+  int64_t storage_offset;
 
   // Check and see if the indexing object triggers advanced indexing semantics
 #ifndef TH_REAL_IS_HALF
@@ -1645,7 +1645,7 @@ static int THPTensor_(setValue)(THPTensor *self, PyObject *index, PyObject *valu
 
   THTensorPtr tresult;
   THStorage *sresult;
-  long storage_offset;
+  int64_t storage_offset;
 
   // Check and see if the indexing object triggers advanced indexing semantics
 #ifndef TH_REAL_IS_HALF
